requirements.txt
networkx
pandas
psutil
numpy
psycopg2-binary
neo4j
pymongo
python-arango
matplotlib
seaborn

